+++ src/model/fonteTensaoControladaCorrente.py
@@ -34,7 +34,7 @@ class FonteTensaoControladaCorrente(ElementoCircuito):
         self.noTensaoNegativo = int(nl[2])
         self.noControlePositivo = int(nl[3])
         self.noControleNegativo = int(nl[4])
-        self.transresistencia = int(nl[5])
+        self.transresistencia = float(nl[5])
         return self
 
     def estampa(
